pyproject.toml
[tool.poetry]
name = "Messenger"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

# telegram
aiogram = "^3.1.1"
aiogram-dialog = "^2.0.0"

# another
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
dependency-injector = "^4.41.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.29"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
